src/modules/slave/task_dealer/docker/segmentation/requirements.txt
numpy==1.26.4
Pillow==10.2.0
